Fix docker build-arg
The ruby Docker image we are using (`ruby:2.3.1-webapp-onbuild`) already specifies the build-args, and it is not possible to override them as they come from an already built image.

The ruby image `ruby:2.5.3-webapp-onbuild` doesn't behave like this, and it allows us to specify those build-args, but this service is an old one and we can't just bump the ruby image version without making sure everything is compatible.

So, as a compromise solution, we are going to rename our build-args so they don't clash with the already defined ones.

Also I've removed the arg `version_number` because it doesn't tell us anything meaninful, it is just the build counter on CircleCI.