Bump version of Ruby
Wheezy and Jessie (except LTS) have been removed from mirrors, so our Dockerfile doesn't work anymore as the Ruby image we were using is too old.

https://lists.debian.org/debian-devel-announce/2019/03/msg00006.html

Bump the version of ruby to 2.5.3 which is conservative enough and should not break anything in this, otherwise, old app, and at the same time add a `docker-compose` file as we did with other service to simplify testing locally.

The bump in ruby version required also a bump in Rubocop and fix the config file to work with the new version.

Prepare the Dockerfile to be able to deploy to Live-1 cluster.