Update CircleCI to work with Live-1 k8s cluster
We had hardcoded the AWS region, and in live-1 this has to change from eu-west-1 to eu-west-2. So better to use the ENV variable already exposed.

Also, for consistency, we are going to name the container in all our applications the same (webapp) as it makes it easier to work with when there are multiple containers (like another one running nginx).