Improve dependencies with aglio
This service uses `aglio` just to transform into a nice looking `html` static documentation an API Blueprint markdown document (api.apib, in the root of the service).

This library depends on lots of other node modules dependencies, some of which produce vuln alerts from Github.
We don't really use this for anything else that generate documentation, so we take out the `yarn.lock` from the repo, add it to the gitignore, and let (as it is now) the docker build produce the documentation using the package.json at build time.

The Dockerfile is a multi-stage docker file, and in the `documentation` stage the dependencies are pulled using the package.json