Periodic maintenance of gems, bump to ruby 2.6.5
We are not yet bumping to Ruby 2.7 but that will come next in a few weeks. It is still a bit new and there might be some incompatibilities so better to wait.

Ruby 2.6.5 is the last release in the 2.6 branch.

Other minor bumps to some other gems, with no breaking changes. Updated config for Puma.