Add missing images and cleanup styles
We were getting 404 errors, and sometimes, when a few of these happened in a short amount of time, an alert triggered in our slack channel.

There were a few missing images. I've added them to the public dir and fixed the routes in the stylesheet. Also I've removed a bunch of other image-related styles as these are not used in this service at all.