Reduce number of 404 errors
We are still getting from time to time what seems to be automated scans in search for known products like Umbraco or Wordpress.

These scans produce quite a few 404 in a very short time span (1 or 2 seconds) and sometimes make our alerts trigger.

I've noticed some genuine 404 errors can be removed:

1. Adding a favicon.ico file (the standard MOJ one)
2. As per MOJ Security Guidance [1], implementing a `/.well_known/security.txt` location.

This is not going to take away all the 404 but help to reduce them. If this continues the other option will be to tweak the alerts.

[1] https://ministryofjustice.github.io/security-guidance/contact/implement-security-txt/#implementing-securitytxt