Bump Sinatra and Grape, and use `dry-validation`
Replace `hanami-validations` with `dry-validation` as otherwise with the bump of Sinatra and Grape, the version of `hanami-validations` that we could use is way outdated.

The bumps in the gems is motivated by the necessity to try to use a more updated Rack version, as the one we currently have is very old and contains some vulnerabilities.