Add headings memoization optimisation
Incidentally as part of the refactor for the validations in PR #106, as I had to run many many tests and many uploads as well, I found the main cause of the veeeery slow file processing times.

The `#processed_headings` method was being called (in the `#transform_mediators` loop) more than 1k times, 1 for each row in the file, and each time it was called it performed the exact same operations on the exact same data, meaning it was a candidate for memoization.

Benchmark results before and after memoization for the exact same file and number of rows (1103):

Before:
```
44.086576   0.220200  44.306776 ( 44.318976)
```

After:
```
0.753861   0.043804   0.797665 (  0.797686)
```