Add ECR images cleaner workflow
This workflow will run on a schedule every Sunday at 9am, as well as it is possible to run it manually (once it is merged to `main` branch).

It will delete images older than 30 days, that are not part of the replica set history for production namespace, but keeping a buffer of 25 images no matter how old they are.